#2351: Move plist parsers to default parse package
