+++ iped-app/resources/config/conf/ParserConfig.xml
@@ -23,8 +23,8 @@
         <parser class="iped.parsers.external.CompositeExternalParser"></parser>
 
         <parser class="org.apache.tika.parser.apple.AppleSingleFileParser"></parser>
-        <parser class="iped.parsers.plist.parser.PListParser"></parser>
-        <parser class="iped.parsers.plist.parser.NSKeyedArchiverParser"></parser>
+        <parser class="iped.parsers.plist.PListParser"></parser>
+        <parser class="iped.parsers.plist.NSKeyedArchiverParser"></parser>
         <!-- <parser class="org.apache.tika.parser.apple.PListParser"></parser> -->
         <parser class="org.apache.tika.parser.asm.ClassParser"></parser>
         <parser class="org.apache.tika.parser.audio.AudioParser"></parser>
